// src/app/Models/FootballClub.ts
export interface IFootballClub{
  regNoOfClub: number;
  coachName: string;
  clubName: string;
  streetOfTheClub: string;
  cityOfTheClub: string;
  postCodeOfTheStreet: number;
  noOfMatchesPlayed: number;
  noOfMatchesWon: number;
  noOfMatchesLost: number;
  noOfMatchesDraw: number;
  noOwnGoalReceived: number;
  noOfGoalScored: number;
  noOfPoints: number;
  goalDiff: number;
}